# backend/docs/swagger.yaml
basePath: /
definitions:
  echo.HTTPError:
    properties:
      message: {}
    type: object
  github_com_kldd0_travel-hack-2024_internal_entity.City:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  github_com_kldd0_travel-hack-2024_internal_entity.Image:
    properties:
      caption:
        type: string
      url:
        type: string
    type: object
  github_com_kldd0_travel-hack-2024_internal_entity.Review:
    properties:
      author:
        description: автор отзыва
        type: string
      frequency:
        description: частота посещения
        type: string
      liked:
        description: рейтинг
        items:
          type: string
        type: array
      local_resident:
        description: местный?
        type: boolean
      negative:
        description: недостатки
        type: string
      positive:
        description: достоинства
        type: string
      tour_id:
        description: id тура
        type: integer
      type:
        description: тип отдыха (в одиночку. с детьми...)
        type: string
      video:
        description: ссылка на видео
        type: string
    type: object
  github_com_kldd0_travel-hack-2024_internal_entity.SimplifiedTourView:
    properties:
      category:
        type: string
      id:
        type: integer
      location:
        type: string
      night_count:
        type: integer
      rating:
        type: integer
      title:
        type: string
      type:
        items:
          $ref: '#/definitions/github_com_kldd0_travel-hack-2024_internal_entity.Tag'
        type: array
    type: object
  github_com_kldd0_travel-hack-2024_internal_entity.Tag:
    properties:
      title:
        type: string
    type: object
  github_com_kldd0_travel-hack-2024_internal_entity.Tour:
    properties:
      category:
        type: string
      comfort_level:
        type: string
      description:
        type: string
      difficulty_level:
        type: string
      faq:
        type: string
      id:
        type: integer
      important_info:
        type: string
      included:
        items:
          type: string
        type: array
      location:
        type: string
      media:
        items:
          $ref: '#/definitions/github_com_kldd0_travel-hack-2024_internal_entity.Image'
        type: array
      nearest_date:
        description: must be updated every req
        type: string
      night_count:
        type: integer
      not_included:
        items:
          type: string
        type: array
      program:
        items:
          type: string
        type: array
      rating:
        type: integer
      reviews:
        description: массиов отзывов
        items:
          $ref: '#/definitions/github_com_kldd0_travel-hack-2024_internal_entity.Review'
        type: array
      title:
        type: string
      tour_dates:
        description: must be updated every req
        items:
          $ref: '#/definitions/github_com_kldd0_travel-hack-2024_internal_entity.TourDate'
        type: array
      type:
        items:
          $ref: '#/definitions/github_com_kldd0_travel-hack-2024_internal_entity.Tag'
        type: array
    type: object
  github_com_kldd0_travel-hack-2024_internal_entity.TourDate:
    properties:
      end:
        type: string
      start:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Additional service for RUSSPASS
  title: Tour Management Service
  version: "1.0"
paths:
  /api/v1/cities/{prefix}:
    get:
      description: Get city by prefix
      parameters:
      - description: City name prefix
        in: path
        name: prefix
        required: true
        type: string
      - description: Output limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_kldd0_travel-hack-2024_internal_entity.City'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Get city by prefix
      tags:
      - cities
  /api/v1/tours:
    get:
      description: Get tours according to search and filter params
      parameters:
      - description: From location name
        in: query
        name: fromName
        type: string
      - description: To location name
        in: query
        name: toName
        type: string
      - description: Date of the tour
        in: query
        name: whenDate
        type: string
      - description: Number of nights in the tour
        in: query
        name: nightsCnt
        type: integer
      - description: Number of adults
        in: query
        name: adults
        type: integer
      - description: Number of children
        in: query
        name: childrens
        type: integer
      - description: Type of the tour
        in: query
        name: tourType
        type: string
      - description: Minimum price
        in: query
        name: priceFrom
        type: string
      - description: Maximum price
        in: query
        name: priceTo
        type: string
      - description: Minimum rating
        in: query
        name: rating
        type: integer
      - description: Guaranteed availability
        in: query
        name: guaranteed
        type: boolean
      - description: Flight is included
        in: query
        name: withFlight
        type: boolean
      - description: Accomodation is included
        in: query
        name: withAcc
        type: boolean
      - description: Nutrition is included
        in: query
        name: withFood
        type: boolean
      - description: The tour takes place on a weekend
        in: query
        name: dayOff
        type: boolean
      - description: Low cost tour
        in: query
        name: lowCost
        type: boolean
      - description: Age group ID
        in: query
        name: ageGroupId
        type: integer
      - description: Tour difficulty level
        in: query
        name: difficulty
        type: integer
      - description: Comfort level
        in: query
        name: comfort
        type: integer
      - description: Food ID
        in: query
        name: foodId
        type: integer
      produces:
      - application/json
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_kldd0_travel-hack-2024_internal_entity.SimplifiedTourView'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Get tours according to search and filter params
      tags:
      - tours
  /api/v1/tours/{id}:
    get:
      description: Get tour by id
      parameters:
      - description: Tour id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_kldd0_travel-hack-2024_internal_entity.Tour'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Get tour by id
      tags:
      - tours
securityDefinitions:
  JWT:
    description: JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"